server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-genai>=1.70.0
